# packages/api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: Ámbar Core API
  version: 0.1.0
  summary: API del core bancario Ámbar (ledger, transferencias, dispositivos, avisos y fraude).
  description: |
    Contrato fuente de verdad del core bancario Ámbar. Todo el código cliente
    (apps móvil y web) se genera desde este archivo y QA lo usa para el mock server.

    Convenciones:
    - Montos en **centavos** como enteros (`125050` = $1,250.50 MXN). Nunca decimales.
    - Toda operación que mueve dinero exige el header `Idempotency-Key` (UUID).
      Repetir la misma llave con el mismo cuerpo devuelve la respuesta original
      con `Idempotent-Replayed: true`; con un cuerpo distinto devuelve 409.
    - Errores en formato RFC 9457 (`application/problem+json`) con un `code` estable.
    - Un recurso de otro usuario responde 404, no 403, para no revelar que existe.
  contact:
    name: José Carlos Bernal Bautista
    email: josecarlosbernal09@gmail.com
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:3000
    description: Local
  - url: https://api.staging.ambar.example
    description: Staging (datos sintéticos)
...
security:
  - cognito: []
paths:
...
  /v1/assistant/messages:
    post:
      tags: [Asistente]
      operationId: sendAssistantMessage
      summary: Envía una pregunta al asistente y recibe su respuesta
      description: |
        Sin `conversation_id` empieza una conversación nueva (dura 24 horas). El asistente puede
        consultar cuentas, movimientos, un resumen del mes y avisos; no puede hacer operaciones.
        Antes de procesar el mensaje se ocultan tarjetas, CLABE, NIP, contraseñas y otros datos
        personales, y `notices` le explica al usuario qué se ocultó. El token debe tener también
        `accounts.read`: el asistente lo usa tal cual para consultar esta API.
      security:
        - cognito: [assistant.chat, accounts.read]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AssistantMessageRequest' }
      responses:
        '200':
          description: Respuesta del asistente
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AssistantReply' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /v1/assistant/conversations/{conversationId}:
    parameters:
      - $ref: '#/components/parameters/ConversationId'
    get:
      tags: [Asistente]
      operationId: getAssistantConversation
      summary: Historial de una conversación propia (texto ya redactado)
      description: Solo las conversaciones del usuario autenticado y que no han vencido.
      security:
        - cognito: [assistant.chat]
      responses:
        '200':
          description: Conversación
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AssistantConversation' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Asistente]
      operationId: deleteAssistantConversation
      summary: Borra una conversación propia
      description: El historial se elimina de inmediato. La auditoría no contiene el texto de la conversación.
      security:
        - cognito: [assistant.chat]
      responses:
        '204':
          description: Conversación borrada
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    cognito:
      type: oauth2
      description: |
        Amazon Cognito (Authorization Code + PKCE). Los scopes se emiten con el
        prefijo del resource server, p. ej. `ambar-api/transfers.write`.
      flows:
        authorizationCode:
          authorizationUrl: https://auth.ambar.example/oauth2/authorize
          tokenUrl: https://auth.ambar.example/oauth2/token
          refreshUrl: https://auth.ambar.example/oauth2/token
          scopes:
            accounts.read: Consultar cuentas y movimientos propios
            accounts.write: Abrir cuentas
            transfers.write: Enviar transferencias
            devices.write: Registrar y revocar dispositivos de la app móvil
            ledger.admin: Operación del ledger (consultar y reversar asientos)
            qa.write: Simular movimientos en entornos no productivos
            assistant.chat: Conversar con el asistente financiero
    cognitoM2M:
      type: oauth2
      description: Cognito client_credentials para la automatización (n8n). Solo dentro de la VPC.
      flows:
        clientCredentials:
          tokenUrl: https://auth.ambar.example/oauth2/token
          scopes:
            internal.notify: Crear avisos para clientes
            internal.fraud: Abrir casos de posible fraude
            internal.reconcile: Consultar la conciliación del ledger y del outbox
  parameters:
...
    ConversationId:
      name: conversationId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
    Uuid:
      type: string
      format: uuid
...
    Timestamp:
      type: string
      format: date-time
...
    AssistantMessageRequest:
      type: object
      additionalProperties: false
      required: [message]
      properties:
        conversation_id: { $ref: '#/components/schemas/Uuid' }
        message:
          type: string
          minLength: 1
          maxLength: 1000
          examples: ['¿En qué gasté más este mes?']
    AssistantReply:
      type: object
      additionalProperties: false
      required: [conversation_id, reply, tools_used, notices]
      properties:
        conversation_id: { $ref: '#/components/schemas/Uuid' }
        reply:
          type: object
          additionalProperties: false
          required: [text, created_at]
          properties:
            text: { type: string, maxLength: 1200 }
            created_at: { $ref: '#/components/schemas/Timestamp' }
        tools_used:
          type: array
          description: Consultas que hizo el asistente para responder (transparencia para el usuario).
          items:
            type: string
            enum: [consultar_cuentas, consultar_movimientos, resumen_del_mes, consultar_avisos]
        notices:
          type: array
          description: Avisos de privacidad, p. ej. que se ocultó una tarjeta del mensaje.
          items: { type: string }
    AssistantConversation:
      type: object
      additionalProperties: false
      required: [id, messages, created_at, expires_at]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        messages:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [role, text, created_at]
            properties:
              role: { type: string, enum: [user, assistant] }
              text: { type: string }
              created_at: { $ref: '#/components/schemas/Timestamp' }
        created_at: { $ref: '#/components/schemas/Timestamp' }
        expires_at: { $ref: '#/components/schemas/Timestamp' }
    Problem:
      type: object
      description: RFC 9457 Problem Details.
      required: [type, title, status, code]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code:
          type: string
          description: |
            Código estable para clientes y QA. Valores actuales:
            VALIDATION_ERROR, UNAUTHENTICATED, FORBIDDEN, NOT_FOUND,
            INSUFFICIENT_FUNDS, ACCOUNT_NOT_ACTIVE, SAME_ACCOUNT,
            IDEMPOTENCY_KEY_REUSED, ALREADY_REVERSED, NOT_REVERSIBLE,
            STEP_UP_REQUIRED, DEVICE_LIMIT_REACHED, FRAUD_CASE_CLOSED,
            RATE_LIMITED, CONVERSATION_FULL, CONVERSATION_BUSY, ASSISTANT_UNAVAILABLE.
        errors:
          type: array
          items:
            type: object
            required: [path, message]
            properties:
              path: { type: string }
              message: { type: string }
        step_up:
          type: object
          description: Solo con code STEP_UP_REQUIRED.
          required: [reason]
          properties:
            reason:
              type: string
              enum: [AMOUNT_THRESHOLD, NEW_BENEFICIARY, INVALID_PROOF]
            threshold:
              type: integer
              format: int64
  responses:
    ValidationError:
      description: La solicitud no cumple el contrato
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthenticated:
      description: Falta el token o es inválido
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    Forbidden:
      description: El token no tiene el scope requerido
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: El recurso no existe o no pertenece al usuario
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    Conflict:
      description: El recurso está en un estado que no admite esta operación
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    RateLimited:
      description: Se alcanzó un límite de uso
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unavailable:
      description: Una dependencia (p. ej. el modelo de IA) no está disponible
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...